fix: checks added for infraConfig table containing missing value or not
